Build the full path of a new file based on its actual parent directory
This partially fixes a bug where all files paths were built relative to the
original parent directory, not the actual directory of a new file.

The bug fix only works for files in directories which existed when the app
was started. The app currently crashes if you add a file to a newly created
directory.