Handle API update errors in FileWatcher
Log errors and ignore, in case there is a temporary issue like the API not being
available.

This is a very naive way to handle the errors, because failed updates will never be
retried. We may want to revisit this if, for example, it would be useful to developers
if updates are retried when the API becomes available again.