Extract file ID from the end of the filename rather than the full filename
S3Ninja does not store files on disk in a nested folder structure. If you upload
a file to S3Ninja with nested folders, e.g. "foo/bar/baz.txt", this is stored in
S3Ninja as "foo_bar_baz.txt".

This means that to extract the file ID (which is a UUID) from an S3Ninja filename,
we have to take the last 32 characters of the filename.

Ignore and log any files which do not fit this pattern.