Improve error message for failed filename match
The filename needs to end in a UUID because S3 ninja stores S3 "folder"
names within the the filename itself. For example, if you try to upload
a file to S3 at the path "foo/bar/baz.txt", S3 ninja stores it on disk
with the filename "foo_bar_baz.txt".