Save fake checksum based on original filename
If the filename starts with "test-checksum-", use the rest of the filename
(ignoring the file extension) as the checksum.

For example, a file with the name "test-checksum-abcde.txt" will generate
a fake checksum "abcde".

Use "fake-checksum" as the default for all other files. Using obvious test
data will make it less likely that we will confuse the value with a real
checksum.

When the checksum check expands in future to compare the client-side and
server-side checksum, this fake check will have to be updated, but this
simple value will work for now.